Oc framework: move classes to proper package.
