+++ src/main/java/br/com/schumaker/octopus/framework/ioc/managed/ManagedClass.java
@@ -1,4 +1,4 @@
-package br.com.schumaker.octopus.framework.ioc;
+package br.com.schumaker.octopus.framework.ioc.managed;
 
 public interface ManagedClass<T> {
