+++ src/main/java/br/com/schumaker/octopus/framework/ioc/managed/ManagedRepository.java
@@ -1,4 +1,4 @@
-package br.com.schumaker.octopus.framework.ioc;
+package br.com.schumaker.octopus.framework.ioc.managed;
 
 import br.com.schumaker.octopus.framework.reflection.ClassReflection;
